# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Console API Documentation",
        "description": "Console Document Management System API - A comprehensive RAG (Retrieval-Augmented Generation) system for managing documents with hierarchical directory structure",
        "contact": {
            "name": "Mauritius Telecom Innovation Team",
            "email": "innovation@telecom.mu"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Console API Server"
        }
    ],
    "paths": {
...
        "/directories": {
            "get": {
                "tags": [
                    "Directories"
                ],
                "summary": "Get directory tree",
                "description": "Retrieve hierarchical directory structure with nested children",
                "operationId": "2bff3cdbef91fa5a1e08614254888f0c",
                "responses": {
                    "200": {
                        "description": "Directory tree retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "name": {
                                                "type": "string"
                                            },
                                            "parent_id": {
                                                "type": "integer",
                                                "nullable": true
                                            },
                                            "created_by": {
                                                "type": "integer"
                                            },
                                            "children": {
                                                "type": "array",
                                                "items": {
                                                    "type": "object"
                                                }
                                            },
                                            "creator": {
                                                "type": "object"
                                            },
                                            "documents": {
                                                "type": "array",
                                                "items": {
                                                    "type": "object"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/directories/{id}/contents": {
            "get": {
                "tags": [
                    "Directories"
                ],
                "summary": "Get directory contents",
                "description": "Get detailed contents of a specific directory including subdirectories and documents",
                "operationId": "99dc28d82392b1f2e8dd03cc441d4f09",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Directory ID (use 0 for root)",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Directory contents retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_directory": {
                                            "type": "object",
                                            "nullable": true
                                        },
                                        "directories": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "documents": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "breadcrumbs": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "stats": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Directory not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}